+++ tests/data_factory.py
@@ -54,7 +54,8 @@ class TestOpener(tools.Opener):
 DEFAULT_METRIC_ALGO_SCRIPT = f"""
 import json
 import substratools as tools
-class TestMetrics(tools.Metrics):
+
+class TestMetrics(tools.MetricAlgo):
     def score(self, inputs, outputs, task_properties):
         y_true = inputs['{InputIdentifiers.datasamples}'][1]
         y_pred = self._get_predictions(inputs['{InputIdentifiers.predictions}'])
@@ -67,7 +68,7 @@ class TestMetrics(tools.Metrics):
             return json.load(f)
 
 if __name__ == '__main__':
-    tools.metrics.execute(TestMetrics())
+    tools.algo.execute(TestMetrics())
 """